Extract nested classes from WinFormsUtils
