+++ src/System.Windows.Forms/src/System/Windows/Forms/TreeNodeCollection.cs
@@ -662,7 +662,7 @@ namespace System.Windows.Forms
 
         public IEnumerator GetEnumerator()
         {
-            return new WindowsFormsUtils.ArraySubsetEnumerator(owner.children, owner.childCount);
+            return new ArraySubsetEnumerator(owner.children, owner.childCount);
         }
     }
 }
