+++ src/System.Windows.Forms/src/System/Windows/Forms/WinFormsUtils.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Collections;
 using System.ComponentModel;
 using System.Diagnostics;
 using System.Drawing;
@@ -14,7 +13,7 @@ using static Interop;
 
 namespace System.Windows.Forms
 {
-    internal sealed class WindowsFormsUtils
+    internal sealed partial class WindowsFormsUtils
     {
         public static readonly ContentAlignment AnyRightAlign = ContentAlignment.TopRight | ContentAlignment.MiddleRight | ContentAlignment.BottomRight;
         public static readonly ContentAlignment AnyLeftAlign = ContentAlignment.TopLeft | ContentAlignment.MiddleLeft | ContentAlignment.BottomLeft;
@@ -382,310 +381,6 @@ namespace System.Windows.Forms
                     return false;
                 }
                 return (remainder >= minValAfterShift && remainder <= maxValAfterShift);
-            }
-        }
-
-        public class ArraySubsetEnumerator : IEnumerator
-        {
-            private readonly object[] _array;
-            private readonly int _total;
-            private int _current;
-
-            public ArraySubsetEnumerator(object[] array, int count)
-            {
-                Debug.Assert(count == 0 || array != null, "if array is null, count should be 0");
-                Debug.Assert(array == null || count <= array.Length, "Trying to enumerate more than the array contains");
-                _array = array;
-                _total = count;
-                _current = -1;
-            }
-
-            public bool MoveNext()
-            {
-                if (_current < _total - 1)
-                {
-                    _current++;
-                    return true;
-                }
-
-                return false;
-            }
-
-            public void Reset()
-            {
-                _current = -1;
-            }
-
-            public object Current => _current == -1 ? null : _array[_current];
-        }
-
-        /// <summary>
-        ///  This is a ControlCollection which can be made readonly. In readonly mode, this
-        ///  ControlCollection throws NotSupportedExceptions for any operation that attempts
-        ///  to modify the collection.
-        /// </summary>
-        internal class ReadOnlyControlCollection : Control.ControlCollection
-        {
-            private readonly bool _isReadOnly;
-
-            public ReadOnlyControlCollection(Control owner, bool isReadOnly) : base(owner)
-            {
-                _isReadOnly = isReadOnly;
-            }
-
-            public override void Add(Control value)
-            {
-                if (IsReadOnly)
-                {
-                    throw new NotSupportedException(SR.ReadonlyControlsCollection);
-                }
-
-                AddInternal(value);
-            }
-
-            internal virtual void AddInternal(Control value) => base.Add(value);
-
-            public override void Clear()
-            {
-                if (IsReadOnly)
-                {
-                    throw new NotSupportedException(SR.ReadonlyControlsCollection);
-                }
-
-                base.Clear();
-            }
-
-            internal virtual void RemoveInternal(Control value) => base.Remove(value);
-
-            public override void RemoveByKey(string key)
-            {
-                if (IsReadOnly)
-                {
-                    throw new NotSupportedException(SR.ReadonlyControlsCollection);
-                }
-
-                base.RemoveByKey(key);
-            }
-
-            public override bool IsReadOnly => _isReadOnly;
-        }
-
-        /// <summary>
-        ///  This control collection only allows a specific type of control
-        ///  into the controls collection. It optionally supports readonlyness.
-        /// </summary>
-        internal class TypedControlCollection : ReadOnlyControlCollection
-        {
-            private readonly Type _typeOfControl;
-            private readonly Control _ownerControl;
-
-            public TypedControlCollection(Control owner, Type typeOfControl, bool isReadOnly) : base(owner, isReadOnly)
-            {
-                _typeOfControl = typeOfControl;
-                _ownerControl = owner;
-            }
-
-            public TypedControlCollection(Control owner, Type typeOfControl) : base(owner, /*isReadOnly*/false)
-            {
-                _typeOfControl = typeOfControl;
-                _ownerControl = owner;
-            }
-
-            public override void Add(Control value)
-            {
-                // Check parenting first for consistency
-                Control.CheckParentingCycle(_ownerControl, value);
-
-                if (value == null)
-                {
-                    throw new ArgumentNullException(nameof(value));
-                }
-                if (IsReadOnly)
-                {
-                    throw new NotSupportedException(SR.ReadonlyControlsCollection);
-                }
-                if (!_typeOfControl.IsAssignableFrom(value.GetType()))
-                {
-                    throw new ArgumentException(string.Format(CultureInfo.CurrentCulture, string.Format(SR.TypedControlCollectionShouldBeOfType, _typeOfControl.Name)), value.GetType().Name);
-                }
-
-                base.Add(value);
-            }
-        }
-
-        /// <summary>
-        ///  DCMapping is used to change the mapping and clip region of the
-        ///  the specified device context to the given bounds. When the
-        ///  DCMapping is disposed, the original mapping and clip rectangle
-        ///  are restored.
-        ///
-        ///  Example:
-        ///
-        ///  using(WindowsFormsUtils.DCMapping mapping = new WindowsFormsUtils.DCMapping(hDC, new Rectangle(10,10, 50, 50) {
-        ///  // inside here the hDC's mapping of (0,0) is inset by (10,10) and
-        ///  // all painting is clipped at (0,0) - (50,50)
-        ///  }
-        ///
-        ///  To use with GDI+ you can get the hDC from the Graphics object. You'd want to do this in a situation where
-        ///  you're handing off a graphics object to someone, and you want the world translated some amount X,Y. This
-        ///  works better than g.TranslateTransform(x,y) - as if someone calls g.GetHdc and does a GDI operation - their
-        ///  world is NOT transformed.
-        ///
-        ///  HandleRef hDC = new HandleRef(this, originalGraphics.GetHdc());
-        ///  try {
-        ///  using(WindowsFormsUtils.DCMapping mapping = new WindowsFormsUtils.DCMapping(hDC, new Rectangle(10,10, 50, 50) {
-        ///
-        ///  // DO NOT ATTEMPT TO USE originalGraphics here - you'll get an Object Busy error
-        ///  // rather ask the mapping object for a graphics object.
-        ///  mapping.Graphics.DrawRectangle(Pens.Black, rect);
-        ///  }
-        ///  }
-        ///  finally { g.ReleaseHdc(hDC);}
-        ///
-        ///  PERF: DCMapping is a structure so that it will allocate on the stack rather than in GC managed
-        ///  memory. This way disposing the object does not force a GC. Since DCMapping objects aren't
-        ///  likely to be passed between functions rather used and disposed in the same one, this reduces
-        ///  overhead.
-        /// </summary>
-        internal struct DCMapping : IDisposable
-        {
-            private DeviceContext _dc;
-            private Graphics _graphics;
-            private Rectangle _translatedBounds;
-
-            public unsafe DCMapping(IntPtr hDC, Rectangle bounds)
-            {
-                if (hDC == IntPtr.Zero)
-                {
-                    throw new ArgumentNullException(nameof(hDC));
-                }
-
-                bool success;
-                IntPtr hOriginalClippingRegion = IntPtr.Zero;
-
-                _translatedBounds = bounds;
-                _graphics = null;
-                _dc = DeviceContext.FromHdc(hDC);
-                _dc.SaveHdc();
-
-                // Retrieve the x-coordinates and y-coordinates of the viewport origin for the specified device context.
-                success = Gdi32.GetViewportOrgEx(hDC, out Point viewportOrg).IsTrue();
-                Debug.Assert(success, "GetViewportOrgEx() failed.");
-
-                // Create a new rectangular clipping region based off of the bounds specified, shifted over by the x & y specified in the viewport origin.
-                IntPtr hClippingRegion = Gdi32.CreateRectRgn(viewportOrg.X + bounds.Left, viewportOrg.Y + bounds.Top, viewportOrg.X + bounds.Right, viewportOrg.Y + bounds.Bottom);
-                Debug.Assert(hClippingRegion != IntPtr.Zero, "CreateRectRgn() failed.");
-
-                try
-                {
-                    // Create an empty region oriented at 0,0 so we can populate it with the original clipping region of the hDC passed in.
-                    hOriginalClippingRegion = Gdi32.CreateRectRgn(0, 0, 0, 0);
-                    Debug.Assert(hOriginalClippingRegion != IntPtr.Zero, "CreateRectRgn() failed.");
-
-                    // Get the clipping region from the hDC: result = {-1 = error, 0 = no region, 1 = success} per MSDN
-                    int result = Gdi32.GetClipRgn(hDC, hOriginalClippingRegion);
-                    Debug.Assert(result != -1, "GetClipRgn() failed.");
-
-                    // Shift the viewpoint origint by coordinates specified in "bounds".
-                    var lastViewPort = new Point();
-                    success = Gdi32.SetViewportOrgEx(hDC, viewportOrg.X + bounds.Left, viewportOrg.Y + bounds.Top, &lastViewPort).IsTrue();
-                    Debug.Assert(success, "SetViewportOrgEx() failed.");
-
-                    RegionType originalRegionType;
-                    if (result != 0)
-                    {
-                        // Get the origninal clipping region so we can determine its type (we'll check later if we've restored the region back properly.)
-                        RECT originalClipRect = new RECT();
-                        originalRegionType = Gdi32.GetRgnBox(hOriginalClippingRegion, ref originalClipRect);
-                        Debug.Assert(originalRegionType != RegionType.ERROR, "ERROR returned from SelectClipRgn while selecting the original clipping region..");
-
-                        if (originalRegionType == RegionType.SIMPLEREGION)
-                        {
-                            // Find the intersection of our clipping region and the current clipping region (our parent's)
-                            //      Returns a NULLREGION, the two didn't intersect.
-                            //      Returns a SIMPLEREGION, the two intersected
-                            //      Resulting region (stuff that was in hOriginalClippingRegion AND hClippingRegion is placed in hClippingRegion
-                            RegionType combineResult = Gdi32.CombineRgn(hClippingRegion, hClippingRegion, hOriginalClippingRegion, Gdi32.CombineMode.RGN_AND);
-                            Debug.Assert((combineResult == RegionType.SIMPLEREGION) ||
-                                            (combineResult == RegionType.NULLREGION),
-                                            "SIMPLEREGION or NULLREGION expected.");
-                        }
-                    }
-                    else
-                    {
-                        // If there was no clipping region, then the result is a simple region.
-                        // We don't need to keep track of the original now, since it is empty.
-                        Gdi32.DeleteObject(hOriginalClippingRegion);
-                        hOriginalClippingRegion = IntPtr.Zero;
-                        originalRegionType = RegionType.SIMPLEREGION;
-                    }
-
-                    // Select the new clipping region; make sure it's a SIMPLEREGION or NULLREGION
-                    RegionType selectResult = Gdi32.SelectClipRgn(hDC, hClippingRegion);
-                    Debug.Assert((selectResult == RegionType.SIMPLEREGION ||
-                                  selectResult == RegionType.NULLREGION),
-                                  "SIMPLEREGION or NULLLREGION expected.");
-                }
-                catch (Exception ex) when (!ClientUtils.IsSecurityOrCriticalException(ex))
-                {
-                    _dc.RestoreHdc();
-                    _dc.Dispose();
-                }
-                finally
-                {
-                    // Delete the new clipping region, as the clipping region for the HDC is now set
-                    // to this rectangle. Hold on to hOriginalClippingRegion, as we'll need to restore
-                    // it when this object is disposed.
-                    success = Gdi32.DeleteObject(hClippingRegion).IsTrue();
-                    Debug.Assert(success, "DeleteObject(hClippingRegion) failed.");
-
-                    if (hOriginalClippingRegion != IntPtr.Zero)
-                    {
-                        success = Gdi32.DeleteObject(hOriginalClippingRegion).IsTrue();
-                        Debug.Assert(success, "DeleteObject(hOriginalClippingRegion) failed.");
-                    }
-                }
-            }
-
-            public void Dispose()
-            {
-                if (_graphics != null)
-                {
-                    // Reset GDI+ if used.
-                    // we need to dispose the graphics object first, as it will do
-                    // some restoration to the ViewPort and ClipRectangle to restore the hDC to
-                    // the same state it was created in
-                    _graphics.Dispose();
-                    _graphics = null;
-                }
-
-                if (_dc != null)
-                {
-                    // Now properly reset GDI.
-                    _dc.RestoreHdc();
-                    _dc.Dispose();
-                    _dc = null;
-                }
-            }
-
-            /// <summary>
-            ///  Allows you to get the graphics object based off of the translated HDC.
-            ///  Note this will be disposed when the DCMapping object is disposed.
-            /// </summary>
-            public Graphics Graphics
-            {
-                get
-                {
-                    Debug.Assert(_dc != null, "unexpected null dc!");
-
-                    if (_graphics == null)
-                    {
-                        _graphics = Graphics.FromHdcInternal(_dc.Hdc);
-                        _graphics.SetClip(new Rectangle(Point.Empty, _translatedBounds.Size));
-                    }
-
-                    return _graphics;
-                }
             }
         }
     }
